+++ lib/ShiftyRequest/Service/make_edit_attendance_service.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module ShiftyRequest
+  module Service
+    class MakeEditAttendanceService
+    end
+  end
+end
