+++ lib/shifty_request.rb
@@ -5,8 +5,8 @@ require 'dotenv/load'
 require_relative 'ShiftyRequest/Model/attendance'
 
 require_relative 'ShiftyRequest/command'
-require_relative 'ShiftyRequest/api_request'
-require_relative 'ShiftyRequest/edit_attendance'
-require_relative 'ShiftyRequest/load_attendance'
+require_relative 'ShiftyRequest/Request/api_request'
+require_relative 'ShiftyRequest/Request/edit_attendance'
+require_relative 'ShiftyRequest/Request/load_attendance'
 
 ShiftyRequest::Command.run(ARGV)
